refactors flask application to be modular
