+++ docker-compose.override.yml
@@ -16,10 +16,11 @@ services:
       dockerfile: Dockerfile
     environment:
       - FLASK_ENV=development
-      - FLASK_APP=app.py
+      - FLASK_APP=main
       - FLASK_DEBUG=1
+      - PYTHONPATH=/app
     volumes:
       - ./flask:/app
     ports:
       - "5002:5002"
-    command: flask run --host=0.0.0.0 --port=5002
+    command: flask run --host=0.0.0.0 --port=5002 --reload
