--- express/.gitignore
@@ -1,4 +1,0 @@
-node_modules
-.DS_Store
-.env
-uploads/*
