--- express/Dockerfile.dev
@@ -1,13 +1,0 @@
-FROM node:18-alpine
-
-WORKDIR /app
-
-COPY package*.json ./
-
-RUN npm install
-
-COPY . .
-
-EXPOSE 3000
-
-CMD ["npx", "nodemon", "server.js"]
