--- express/middleware/uploadMiddleware.js
@@ -1,16 +1,0 @@
-const multer = require('multer')
-const path = require('path')
-
-const storage = multer.diskStorage({
-  destination: function (req, file, cb) {
-    console.log(__dirname)
-    cb(null, path.join(__dirname, '..', 'uploads'))
-  },
-  filename: function(req, file, cb) {
-    cb(null, `${Date.now()}-${file.originalname}`)
-  }
-})
-
-const upload = multer({ storage })
-
-module.exports = upload
