--- express/package.json
@@ -1,24 +1,0 @@
-{
-  "name": "express",
-  "version": "1.0.0",
-  "main": "server.js",
-  "scripts": {
-    "start": "node server.js",
-    "dev": "nodemon server.js"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "description": "",
-  "dependencies": {
-    "axios": "^1.8.1",
-    "dotenv": "^16.4.7",
-    "express": "^4.21.2",
-    "form-data": "^4.0.2",
-    "mongoose": "^8.10.0",
-    "multer": "^1.4.4"
-  },
-  "devDependencies": {
-    "nodemon": "^3.1.9"
-  }
-}
