--- express/routes/index.js
@@ -1,8 +1,0 @@
-const express = require('express')
-const router = express.Router()
-
-const apiRouter = require('./api')
-
-router.use('/api', apiRouter)
-
-module.exports = router
