+++ flask/Dockerfile
@@ -9,4 +9,4 @@ COPY . .
 
 EXPOSE 5002
 
-CMD ["python", "app.py"]
+CMD ["python", "main.py"]
